structure reorganized
