--- scheduler.py
@@ -1,64 +1,0 @@
-
-class CircularLRBeta:
-
-    def __init__(
-        self, optimizer, lr_max, lr_divider, cut_point, step_size, momentum=None
-    ):
-        self.lr_max = lr_max
-        self.lr_divider = lr_divider
-        self.cut_point = step_size // cut_point
-        self.step_size = step_size
-        self.iteration = 0
-        self.cycle_step = int(step_size * (1 - cut_point / 100) / 2)
-        self.momentum = momentum
-        self.optimizer = optimizer
-
-    def get_lr(self):
-        if self.iteration > 2 * self.cycle_step:
-            cut = (self.iteration - 2 * self.cycle_step) / (
-                self.step_size - 2 * self.cycle_step
-            )
-            lr = self.lr_max * (1 + (cut * (1 - 100) / 100)) / self.lr_divider
-
-        elif self.iteration > self.cycle_step:
-            cut = 1 - (self.iteration - self.cycle_step) / self.cycle_step
-            lr = self.lr_max * (1 + cut * (self.lr_divider - 1)) / self.lr_divider
-
-        else:
-            cut = self.iteration / self.cycle_step
-            lr = self.lr_max * (1 + cut * (self.lr_divider - 1)) / self.lr_divider
-
-        return lr
-
-    def get_momentum(self):
-        if self.iteration > 2 * self.cycle_step:
-            momentum = self.momentum[0]
-
-        elif self.iteration > self.cycle_step:
-            cut = 1 - (self.iteration - self.cycle_step) / self.cycle_step
-            momentum = self.momentum[0] + cut * (self.momentum[1] - self.momentum[0])
-
-        else:
-            cut = self.iteration / self.cycle_step
-            momentum = self.momentum[0] + cut * (self.momentum[1] - self.momentum[0])
-
-        return momentum
-
-    def step(self):
-        lr = self.get_lr()
-
-        if self.momentum is not None:
-            momentum = self.get_momentum()
-
-        self.iteration += 1
-
-        if self.iteration == self.step_size:
-            self.iteration = 0
-
-        for group in self.optimizer.param_groups:
-            group['lr'] = lr
-
-            if self.momentum is not None:
-                group['betas'] = (momentum, group['betas'][1])
-
-        return lr
