--- setup.cfg
@@ -1,6 +1,0 @@
-[flake8]
-# Default of black
-max-line-length = 88
-# Ignore whitespace before : (Black formats non-trivial slicing with spaces)
-# and line break before binary operator (same reason)
-ignore= E203,W503
