+++ src/configs/SATRN.yaml
@@ -1,0 +1,39 @@
+network:
+  enc: "Transformer"  # Transformer, CRNN
+  dec: "Transformer"  # Transformer, Attention, CRNN
+input_size:
+  height: 128
+  width: 128
+dec_layers: 3
+rgb: 1
+src_dim: 300
+encoder_dim: 300
+filter_size: 600
+encoder_layers: 6
+checkpoint: ""
+prefix: ""
+
+data: # now supports only one
+  gt_paths:
+    - "../data/Aida/images/train.tsv"  # for now, train
+    - "../data/Aida/images/validation.tsv"  # for now, valid
+  token_paths: "../data/Aida/tokens.txt"  # TODO: to list
+  dataset_proportions:
+    - 1.0
+  split_proportions:
+    train: 0.8
+    valid: 0.2
+  crop: True
+  
+batch_size: 24
+num_workers: 4
+num_epochs: 30
+print_epochs: 1
+dropout_rate: 0.1
+teacher_forcing_ratio: 0.5
+max_grad_norm: 2.0
+seed: 1234
+optimizer:
+  optimizer: 'Adam' # Adam, Adadelta
+  lr: 5e-4 # 1e-4
+  weight_decay: 1e-4
